Update imgaussfiltTest.cpp
Update imgaussfiltTest.cpp

+++ test/imgaussfiltTest.cpp
@@ -17,7 +17,7 @@ void imgaussfiltTest(std::string_view input_image_path = "InputImages/1", std::s
     for(int sigma = 1; sigma < 10; ++sigma)
     {
         auto output_img = TinyDIP::im2uint8(
-                                TinyDIP::imgaussfilt(TinyDIP::im2double(input_img), sigma)
+                                TinyDIP::imgaussfilt(TinyDIP::im2double(input_img), static_cast<double>(sigma))
                                 );
         TinyDIP::bmp_write(
             (std::string(output_image_path) + std::string("_sigma=") + std::to_string(sigma)).c_str(),
